register form completo

+++ database/migrations/2023_02_16_153310_create_dishes_table.php
@@ -16,7 +16,7 @@ class CreateDishesTable extends Migration
         Schema::create('dishes', function (Blueprint $table) {
             $table->id();
             $table->string('name', 100);
-            $table->decimal('price');
+            $table->unsignedInteger('price');
             $table->string('image', 200);
             $table->boolean('visibility')->default(0);
             $table->text('description');
